rename `JavaType` -> `JavaClassDescriptor`
We need to free the name `JavaType` if we want to add support for generics to be consistent with the Reflection API. I.e. if the Reflection API provides a super type `Type` for `Class` and `TypeVariable`, we should probably mimic this as `JavaType`, `JavaClass` and `JavaTypeVariable`. Otherwise we break the consistency with the Reflection API there, making the API confusing. Note that I do not consider `JavaClassDescriptor` the best name in the universe, but merely could not come up with anything better. It is pretty much all the information we can parse out of the info parsed by ASM, name parts, component type, etc.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>