generify type argument processing
Reduced the number of visitor objects needed and generalize the design. We will need to apply more recursion to be able to follow the tree of arbitrarily nested types, e.g. `Foo<Bar<Baz<? extends Quux<T>, ?>, ?>, T>`. It will be possible to recursively reuse the type argument processors in the next step.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>